+++ bin/json5_model.dart
@@ -94,18 +94,20 @@ class %s {
 
 """;
 
-const topListTpl = """List<%s> %sListFormJson(List<Map<String, dynamic>> json) =>
+const topListTpl =
+    """List<%s> %sListFormJson(List<Map<String, dynamic>> json) =>
     json.map((e) => %s.fromJson(e)).toList();
 
 List<Map<String, dynamic>> %sListToJson(List<%s> instance) =>
     instance.map((e) => e.toJson()).toList();""";
 
-void main(List<String> args) {
+Future<void> main(List<String> args) async {
   String? src;
   String? dist;
   String? tag;
   bool nullable = false;
   bool clean = false;
+  bool once = true;
   final parser = ArgParser();
   parser.addOption(
     'src',
@@ -119,7 +121,12 @@ void main(List<String> args) {
     callback: (v) => dist = v,
     help: "Specify the dist directory.",
   );
-
+  parser.addOption(
+    'once',
+    defaultsTo: "true",
+    callback: (v) => once = (v != 'false'),
+    help: "Automatically rename the JSON file to start with '_' after build",
+  );
   parser.addOption(
     'tag',
     defaultsTo: '\$',
@@ -134,17 +141,33 @@ void main(List<String> args) {
   // return;
   if (clean) {
     br.run(['clean']);
-  } else if (generateModelClass(src!, dist!, tag!, nullable: nullable)) {
-    br.run(['build', '--delete-conflicting-outputs']);
+  } else {
+    var files = <String>[];
+    final isGenerated = generateModelClass(src!, dist!, tag!,
+        nullable: nullable, once: once, onGenerate: (f) { files = f;});
+    if (isGenerated) {
+      await br.run(['build', '--delete-conflicting-outputs']);
+      final all = files.map((e) => path.join(path.dirname(e), '*'));
+      Process.run('git', [
+        'add',
+        ...all
+      ]).then((ProcessResult result) {
+        if (result.exitCode == 0) {
+          print('git add succeed');
+        } else {
+          print('git add fail');
+        }
+      }).catchError((error) {
+        print('git add fail: $error');
+      });
+    }
   }
 }
 
-bool generateModelClass(
-  String srcDir,
-  String distDir,
-  String tag, {
-  required bool nullable,
-}) {
+bool generateModelClass(String srcDir, String distDir, String tag,
+    {required bool nullable,
+    required bool once,
+    Function(List<String> allGeneratedFiles)? onGenerate}) {
   if (srcDir.endsWith("/")) srcDir = srcDir.substring(0, srcDir.length - 1);
   if (distDir.endsWith("/")) distDir = distDir.substring(0, distDir.length - 1);
 
@@ -157,6 +180,8 @@ bool generateModelClass(
   }
 
   File file;
+
+  List<String> allGeneratedFiles = List.empty(growable: true);
 
   fileList.forEach((f) {
     if (FileSystemEntity.isFileSync(f.path)) {
@@ -171,8 +196,12 @@ bool generateModelClass(
       if ((!isJson && !isJson5) || fileName.startsWith("_")) return;
 
       final relative = path.relative(f.path, from: srcDir);
-      final dartFilePath =
+      var dartFilePath =
           path.join(distDir, path.setExtension(relative, ".dart"));
+      dartFilePath = path.join(
+          path.dirname(dartFilePath),
+          path.basenameWithoutExtension(dartFilePath),
+          path.basename(dartFilePath));
 
       final json = json5Decode(file.readAsStringSync());
       //generated class name
@@ -278,6 +307,29 @@ bool generateModelClass(
       File(dartFilePath)
         ..createSync(recursive: true)
         ..writeAsStringSync(dist);
+
+      var sourceFilePath = file.path;
+      if (once) {
+        final newName = "_" + path.basename(file.path);
+        sourceFilePath = path.join(file.parent.path, newName);
+        file.renameSync(sourceFilePath);
+      }
+
+      allGeneratedFiles.add(dartFilePath);
+
+      Process.run('git', [
+        'add',
+        sourceFilePath
+      ]).then((ProcessResult result) {
+        if (result.exitCode == 0) {
+          print('git add succeed');
+        } else {
+          print('git add fail');
+        }
+      }).catchError((error) {
+        print('git add fail: $error');
+      });
+
       // indexFile = exportIndexFile(
       //     className, json is List, dartFilePath, distDir, indexFile);
       // print('done: ${f.path} -> $dartFilePath');
@@ -290,6 +342,7 @@ bool generateModelClass(
   // }
   print("src directory：" + path.canonicalize(srcDir));
   print("dist directory：" + path.canonicalize(distDir));
+  onGenerate?.call(allGeneratedFiles);
   return true;
 }